requirements.txt
Flask==3.0.0
pandas==2.1.4
python-dateutil==2.8.2
openpyxl==3.1.2
Werkzeug==3.0.1
gunicorn==21.2.0 